markit-backend/markit-proj/requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.0
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-celery-beat==1.5.0
django-cors-headers==3.2.0
django-model-utils==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-stdimage==5.0.1
django-timezone-field==4.0
djangorestframework==3.10.3
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.3
jsonfield==2.0.2
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==6.2.1
progressbar2==3.47.0
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-crontab==2.4.0
python-dateutil==2.8.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
service-identity==18.1.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.8.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
